xiaozhi-esp32-server-main/main/xiaozhi-server/requirements_uv.txt
aioconsole==0.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.7.0
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
apscheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
azure-ai-documentintelligence==1.0.2
azure-core==1.34.0
azure-identity==1.22.0
baidu-aip==4.16.13
beautifulsoup4==4.13.3
black==25.1.0
bs4==0.0.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cnlunar==0.2.0
cobble==0.1.4
cozepy==0.12.0
crcmod==1.7
cryptography==44.0.3
dashscope==1.23.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
edge-tts==7.0.2
editdistance==0.8.1
et-xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.18.0
flake8==7.3.0
frozenlist==1.5.0
fsspec==2025.3.2
funasr==1.2.3
getmac==0.9.5
google-api-core==2.24.2
google-auth==2.38.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hf-xet==1.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.31.1
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
isort==6.0.1
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
jinja2==3.1.6
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
kaldiio==2.18.1
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
lxml==5.4.0
mammoth==1.9.0
markdownify==1.1.0
markitdown==0.0.2
markupsafe==3.0.2
mcp==1.6.0
mem0ai==0.1.62
memobase==0.0.23
modelscope==1.25.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.3.2
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
omegaconf==2.3.0
openai==1.78.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opuslib-next==1.1.2
oss2==2.19.1
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pathvalidate==3.2.3
pdfminer-six==20250506
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
portalocker==3.1.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
puremagic==1.29
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pydub==0.25.1
pyjwt==2.8.0
pynndescent==0.5.13
pyparsing==3.2.3
pysocks==1.7.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-pptx==1.0.2
pytorch-wpe==0.0.1
pytz==2025.2
pyyaml==6.0.2
pyyml==0.0.2
requests==2.32.3
rsa==4.9
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==5.1.0
sentencepiece==0.2.0
sherpa-onnx==1.12.9
silero-vad==5.1.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
speechrecognition==3.14.2
sqlalchemy==2.0.43
srt==3.5.3
sse-starlette==2.2.1
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tensorboardx==2.6.2.2
threadpoolctl==3.6.0
torch==2.6.0
torch-complex==0.4.4
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.55.2
triton==3.2.0
typing-extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
umap-learn==0.5.7
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
win32-setctime==1.2.0
xlrd==2.0.1
xlsxwriter==3.2.3
yarl==1.19.0
youtube-transcript-api==1.0.3
